storage(rpc): move kv impls to storage crate
